scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
IMDbPY>=2022.7.9

